Move view component JS to frontend folder
This is to make it consistent with what we've done for CSS, where we
no longer use the "sidecar" pattern, but instead have it alongside
other assets of the same kind.
