+++ app/components/README.md
@@ -2,22 +2,26 @@
 JS and SCSS files are optional
 ```
 /app
+  /assets
+    /stylesheets
+      /components
+        my_widget_component.scss
   /components
     /my_widget_component
-      my_widget_component.js
-      my_widget_component.test.js
-      my_widget_component.scss
       my_widget_component.html.slim
     my_widget_component.rb
+  /frontend
+    /src
+      /components
+        /my_widget_component
+          my_widget_component.js
+          my_widget_component.test.js
 ```
 
 The `_component` postfix is integral to how view components work and is essential to include this.
 
 ### Slim usage
 `= render(MyWidgetComponent.new()`
-
-## JS and styles
-The script `frontend/src/components.js` recursivley looks in the `app/components` folder for any Javascript or SASS files and adds these to the Webpack pipeline that builds the application bundle.
 
 ### SCSS usage
 Component SASS files should import `base_component.scss`. This currently makes the settings, tools and helpers of GovUK-frontend available so the familiar mixins and variables can be used e.g `$govuk-border-colour` or `$govuk-border-colour`. It also makes it bit more flexible and easy if other dependencies are needed in future.
