+++ app/frontend/packs/application.js
@@ -5,7 +5,6 @@ import Rails from 'rails-ujs';
 import { initAll } from 'govuk-frontend';
 
 import 'src/application';
-import 'src/components';
 
 Rails.start();
