+++ app/frontend/src/application.js
@@ -6,21 +6,21 @@ import '@stimulus/polyfills';
 import { Application } from '@hotwired/stimulus';
 
 // view components
-import FiltersController from '../../components/filters_component/filters_component';
-import MapController from '../../components/map_component/map_component';
-import MapSidebarController from '../../components/map_component/sidebar';
+import FiltersController from './components/filters_component/filters_component';
+import MapController from './components/map_component/map_component';
+import MapSidebarController from './components/map_component/sidebar';
 
 // js components
-import AutocompleteController from './components/autocomplete/autocomplete';
-import ClipboardController from './components/clipboard/clipboard';
-import FormController from './components/form/form';
-import LocationFinderController from './components/locationFinder/locationFinder';
-import ManageQualificationsController from './components/manageQualifications/manageQualifications';
-import PanelController from './components/panel/panel';
-import ShowHiddenContentController from './components/showHiddenContent/showHiddenContent';
-import TrackedLinkController from './components/trackedLink/trackedLink';
-import UploadDocumentsController from './components/uploadDocuments/uploadDocuments';
-import UtilsController from './components/utils';
+import AutocompleteController from './js_components/autocomplete/autocomplete';
+import ClipboardController from './js_components/clipboard/clipboard';
+import FormController from './js_components/form/form';
+import LocationFinderController from './js_components/locationFinder/locationFinder';
+import ManageQualificationsController from './js_components/manageQualifications/manageQualifications';
+import PanelController from './js_components/panel/panel';
+import ShowHiddenContentController from './js_components/showHiddenContent/showHiddenContent';
+import TrackedLinkController from './js_components/trackedLink/trackedLink';
+import UploadDocumentsController from './js_components/uploadDocuments/uploadDocuments';
+import UtilsController from './js_components/utils';
 
 Sentry.init({
   // `sentryConfig` is set from the application layout
