--- app/frontend/src/components.js
@@ -1,3 +1,0 @@
-const importAll = (r) => r.keys().forEach(r);
-
-importAll(require.context('../../components', true, /[_/]component\.(js)$/));
