+++ app/frontend/src/components/filters_component/filters_component.js
