+++ app/frontend/src/components/map_component/cluster.js
