+++ app/frontend/src/components/map_component/cluster.test.js
