+++ app/frontend/src/components/map_component/map.js
