+++ app/frontend/src/components/map_component/map.test.js
