+++ app/frontend/src/components/map_component/map_component.js
