+++ app/frontend/src/components/map_component/marker/__mocks__/service.js
