+++ app/frontend/src/components/map_component/marker/service.js
