+++ app/frontend/src/components/map_component/marker/template.js
