+++ app/frontend/src/components/map_component/marker/template.test.js
