+++ app/frontend/src/components/map_component/sidebar.js
