+++ app/frontend/src/components/map_component/sidebar.test.js
