+++ app/frontend/src/js_components/__mocks__/loader.js
