+++ app/frontend/src/js_components/clipboard/clipboard.js
