+++ app/frontend/src/js_components/loadingIndicator/loadingIndicator.js
