+++ app/frontend/src/js_components/panel/panel.js
