+++ app/frontend/src/js_components/panel/panel.test.js
