+++ app/frontend/src/js_components/showHiddenContent/showHiddenContent.test.js
